sidebar refactor, and some question box styling
