+++ src/components/navbar.component.tsx
@@ -12,17 +12,10 @@ import React, { useState, useEffect } from "react";
 import { useHistory } from "react-router";
 import clsx from "clsx";
 import { createStyles, makeStyles, useTheme, Theme } from "@material-ui/core/styles";
-import { Drawer, AppBar, Toolbar, List } from "@material-ui/core";
-import { CssBaseline, Typography, Divider, IconButton } from "@material-ui/core";
-import { ListItem, ListItemIcon, ListItemText, Menu, MenuItem, Box } from "@material-ui/core";
-import MenuIcon from "@material-ui/icons/Menu";
-import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
-import ChevronRightIcon from "@material-ui/icons/ChevronRight";
-import RecordVoiceOverIcon from '@material-ui/icons/RecordVoiceOver';
-import QuestionAnswerIcon from "@material-ui/icons/QuestionAnswer";
+import { AppBar, Toolbar } from "@material-ui/core";
+import { CssBaseline, Typography, IconButton } from "@material-ui/core";
+import { Menu, MenuItem, Box } from "@material-ui/core";
 import AccountCircle from "@material-ui/icons/AccountCircle";
-import LiveHelpIcon from "@material-ui/icons/LiveHelp";
-import HelpOutlineIcon from '@material-ui/icons/HelpOutline';
 import * as loginRemote from '../remotes/login.remote'
 
 
@@ -52,32 +45,6 @@ const useStyles = makeStyles((theme: Theme) =>
     menuButton: {
       marginRight: 36,
     },
-    hide: {
-      display: "none",
-    },
-    drawer: {
-      width: drawerWidth,
-      flexShrink: 0,
-      whiteSpace: "nowrap",
-    },
-    drawerOpen: {
-      width: drawerWidth,
-      transition: theme.transitions.create("width", {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-    },
-    drawerClose: {
-      transition: theme.transitions.create("width", {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-      }),
-      overflowX: "hidden",
-      width: theme.spacing(7) + 1,
-      [theme.breakpoints.up("sm")]: {
-        width: theme.spacing(9) + 1,
-      },
-    },
     toolbar: {
       display: "flex",
       alignItems: "center",
@@ -85,10 +52,6 @@ const useStyles = makeStyles((theme: Theme) =>
       padding: theme.spacing(0, 1),
       // necessary for content to be below app bar
       ...theme.mixins.toolbar,
-    },
-    content: {
-      flexGrow: 1,
-      padding: theme.spacing(3),
     },
     orange: {
       color: "#F26925",
@@ -152,7 +115,6 @@ export const NavbarComponent: React.FC<any> = (props) => {
   const history = useHistory();
   const theme = useTheme();
   const classes = useStyles(props);
-  const [open, setOpen] = React.useState(false);
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
   const [
     mobileMoreAnchorEl,
@@ -160,12 +122,7 @@ export const NavbarComponent: React.FC<any> = (props) => {
   ] = React.useState<null | HTMLElement>(null);
   const isMenuOpen = Boolean(anchorEl);
   const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
-  const handleDrawerClose = () => {
-    setOpen(false);
-  };
+
   const handleProfileMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
     setAnchorEl(event.currentTarget);
   };
@@ -234,22 +191,10 @@ export const NavbarComponent: React.FC<any> = (props) => {
       <CssBaseline />
       <AppBar
         position="fixed"
-        className={clsx(classes.appBar, {
-          [classes.appBarShift]: open,
-        })}
+        className={clsx(classes.appBar)}
       >
         <Toolbar className={classes.arrangement}>
           <Box className={classes.arrangementInternal}>
-            <IconButton
-              aria-label="open drawer"
-              onClick={handleDrawerOpen}
-              edge="start"
-              className={clsx(classes.menuButton, classes.orange, {
-                [classes.hide]: open,
-              })}
-            >
-              <MenuIcon fontSize="large" />
-            </IconButton>
             <Box className={classes.imageDoor}>
               <img
                 src={require("../logo/image.png")}
@@ -278,118 +223,6 @@ export const NavbarComponent: React.FC<any> = (props) => {
           </Box>
         </Toolbar>
       </AppBar>
-      <Drawer
-        variant="permanent"
-        className={clsx(classes.drawer, {
-          [classes.drawerOpen]: open,
-          [classes.drawerClose]: !open,
-        })}
-        classes={{
-          paper: clsx({
-            [classes.drawerOpen]: open,
-            [classes.drawerClose]: !open,
-          }),
-        }}
-      >
-        <div className={classes.toolbar}>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === "rtl" ? (
-              <ChevronRightIcon />
-            ) : (
-                <ChevronLeftIcon />
-              )}
-          </IconButton>
-        </div>
-        <Divider />
-        <List>
-          {["Feed"].map((text, index) => (
-            <ListItem
-              id="feedTab"
-              onClick={() => history.push("/feed")}
-              style={{ color: "#F26925" }}
-              button
-              key={text}
-            >
-              <ListItemIcon>
-                <LiveHelpIcon
-                  onClick={() => history.push("/feed")}
-                  style={{ color: "#F26925" }}
-                />
-              </ListItemIcon>
-              <ListItemText primary={text} />
-            </ListItem>
-          ))}
-          <Divider />
-          {["Post A Question"].map((text, index) => (
-            <ListItem
-              id="postQuestionTab"
-              onClick={() => {
-                history.push("/question");
-              }}
-              style={{ color: "#F26925" }}
-              button
-              key={text}
-            >
-              <ListItemIcon>
-                <QuestionAnswerIcon
-                  onClick={() => {
-                    history.push("/question");
-                  }}
-                  style={{ color: "#F26925" }}
-                />
-              </ListItemIcon>
-              <ListItemText primary={text} />
-            </ListItem>
-          ))}
-          <Divider />
-          {["FAQ"].map((text, index) => (
-            <ListItem
-              onClick={() => {
-                history.push("/faq")
-              }}
-              style={{ color: "#F26925" }}
-              button
-              key={text}
-              >
-                
-
-              <ListItemIcon>
-                <RecordVoiceOverIcon
-                  onClick={() => {
-                    history.push("/faq")}} style={{ color: "#F26925" }}>
-                      
-                </RecordVoiceOverIcon>
-               
-              </ListItemIcon>
-              <ListItemText primary={text} />
-            </ListItem>
-            ))}
-            <Divider />
-              {["Question"].map((text, index) => (
-                <ListItem
-                  id="allQuestionsTab"
-                  onClick={() => {
-                    history.push("/revatureQuestion");
-
-                  }}
-                  style={{ color: "#F26925" }}
-                  button
-                  key={text}
-
-                >
-                  <ListItemIcon>
-                  <HelpOutlineIcon
-                    onClick={() => {
-                      history.push("/revatureQuestion");
-                    }}
-                    style={{ color: "#F26925" }}
-                  />
-                  </ListItemIcon>
-                  <ListItemText primary={text} />
-                </ListItem>
-              ))}
-        </List>
-      </Drawer>
         {renderMobileMenu}
         {renderMenu}
     </div>
